Moved some code to helper functions
