+++ e2e/(template).test.ts
@@ -1,10 +1,11 @@
-/*
-import { test, loginToExtension } from './utils/helper';
-
 // Uncomment and duplicate this test to create a new test for recording
 // Turn off the cleanup in the IDE to maintain the user account to save registering again
+
+/*
+import { test, loginAsTestUser } from './utils/helper';
+
 test('new test', async ({ page, extensionId }) => {
-  await loginToExtension({ page, extensionId });
+  await loginAsTestUser({ page, extensionId });
   await page.pause();
   //
 });
