+++ e2e/login.test.ts
@@ -1,0 +1,5 @@
+import { test, loginAsTestUser } from './utils/helper';
+
+test('Login test', async ({ page, extensionId }) => {
+  await loginAsTestUser({ page, extensionId });
+});
