--- e2e/wallet.test.ts
@@ -1,5 +1,0 @@
-import { test, loginToExtension } from './utils/helper';
-
-test('Login test', async ({ page, extensionId }) => {
-  await loginToExtension({ page, extensionId });
-});
